requirements.txt
flask==2.1.2
flask-restful==0.3.9
flask_sqlalchemy==2.5.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
Faker==14.0.0
requests==2.25.1
requests-cache==0.9.6
kafka-python==2.0.2
gunicorn==20.1.0
flask_cors==3.0.10
flask_jwt_extended==4.4.4
marshmallow-enum==1.5.1
pydub==0.25.1
gevent==22.10.1
prometheus-flask-exporter==0.20.3
wheel
mysqlclient==2.1.1
mysql-connector==2.2.9
google-cloud>=0.27.0
google-cloud-core>=0.27.1
google-cloud-datastore>=1.3.0
google-cloud-storage>=1.4.0
google-cloud-pubsub>=2.13.11